// src/app/models/satilik.ts
export class Satilik {
  keySat: string;
  kayTarihSat: string;
  duzTarihSat: string;
  urunIddSat: number;
  aracAdiSat: string;
  aracMarkSat: number;
  urunKatIddSat: number;
  yakitSat: string;
  vitesSat:string;
  fiyatSat:number;
  baslikSat:string;
  aciklamaSat:string;
  islemSat:boolean;


}