// src/app/models/hurda.ts
export class Hurda {
  keyHur: string;
  kayTarihHur: string;
  duzTarihHur: string;
  urunIddHur: number;
  aracAdiHur: string;
  aracMarkHur: number;
  urunKatIddHur: number;
  yakitHur: string;
  vitesHur:string;
  fiyatHur:number;
  baslikHur:string;
  aciklamaHur:string;
  islemHur:boolean;


}